requirements.txt
torch
torchvision
numpy
matplotlib
transformers
datasets
tqdm
scikit-learn
seaborn
Pillow